+++ include/UDRefl/ReflMngr.h
@@ -361,21 +361,21 @@ namespace Ubpa::UDRefl {
 			StrID methodID,
 			void* result_buffer = nullptr,
 			std::span<const TypeID> argTypeIDs = {},
-			ArgsBuffer args_buffer = nullptr) const;
+			ArgPtrBuffer argptr_buffer = nullptr) const;
 
 		InvokeResult Invoke(
 			ConstObjectPtr obj,
 			StrID methodID,
 			void* result_buffer = nullptr,
 			std::span<const TypeID> argTypeIDs = {},
-			ArgsBuffer args_buffer = nullptr) const;
+			ArgPtrBuffer argptr_buffer = nullptr) const;
 
 		InvokeResult Invoke(
 			ObjectPtr obj,
 			StrID methodID,
 			void* result_buffer = nullptr,
 			std::span<const TypeID> argTypeIDs = {},
-			ArgsBuffer args_buffer = nullptr) const;
+			ArgPtrBuffer argptr_buffer = nullptr) const;
 
 		// -- template --
 
@@ -387,11 +387,11 @@ namespace Ubpa::UDRefl {
 		InvocableResult IsInvocable      (TypeID typeID, StrID methodID) const;
 
 		template<typename T>
-		T InvokeRet(TypeID      typeID, StrID methodID, std::span<const TypeID> argTypeIDs = {}, ArgsBuffer args_buffer = nullptr) const;
+		T InvokeRet(TypeID      typeID, StrID methodID, std::span<const TypeID> argTypeIDs = {}, ArgPtrBuffer argptr_buffer = nullptr) const;
 		template<typename T>
-		T InvokeRet(ConstObjectPtr obj, StrID methodID, std::span<const TypeID> argTypeIDs = {}, ArgsBuffer args_buffer = nullptr) const;
+		T InvokeRet(ConstObjectPtr obj, StrID methodID, std::span<const TypeID> argTypeIDs = {}, ArgPtrBuffer argptr_buffer = nullptr) const;
 		template<typename T>
-		T InvokeRet(ObjectPtr      obj, StrID methodID, std::span<const TypeID> argTypeIDs = {}, ArgsBuffer args_buffer = nullptr) const;
+		T InvokeRet(ObjectPtr      obj, StrID methodID, std::span<const TypeID> argTypeIDs = {}, ArgPtrBuffer argptr_buffer = nullptr) const;
 
 		template<typename... Args>
 		InvokeResult InvokeArgs(TypeID      typeID, StrID methodID, void* result_buffer, Args&&... args) const;
@@ -417,8 +417,8 @@ namespace Ubpa::UDRefl {
 		bool IsCopyConstructible(TypeID typeID) const;
 		bool IsMoveConstructible(TypeID typeID) const;
 
-		bool NonArgCopyConstruct(ObjectPtr      obj, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const;
-		bool Construct          (ObjectPtr      obj, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const;
+		bool NonArgCopyConstruct(ObjectPtr      obj, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const;
+		bool Construct          (ObjectPtr      obj, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const;
 		bool Destruct           (ConstObjectPtr obj) const;
 
 		void* Malloc(size_t size) const;
@@ -427,11 +427,11 @@ namespace Ubpa::UDRefl {
 		void* AlignedMalloc(size_t size, size_t alignment) const;
 		bool  AlignedFree  (void* ptr) const;
 
-		ObjectPtr NonArgCopyNew(TypeID      typeID, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const;
-		ObjectPtr New          (TypeID      typeID, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const;
+		ObjectPtr NonArgCopyNew(TypeID      typeID, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const;
+		ObjectPtr New          (TypeID      typeID, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const;
 		bool      Delete       (ConstObjectPtr obj) const;
 
-		SharedObject MakeShared(TypeID typeID, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const;
+		SharedObject MakeShared(TypeID typeID, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const;
 
 		// -- template --
 
@@ -574,21 +574,21 @@ namespace Ubpa::UDRefl {
 			TypeID typeID,
 			StrID methodID,
 			std::span<const TypeID> argTypeIDs = {},
-			ArgsBuffer args_buffer = nullptr,
+			ArgPtrBuffer argptr_buffer = nullptr,
 			std::pmr::memory_resource* result_rsrc = std::pmr::get_default_resource()) const;
 
 		SharedObject MInvoke(
 			ConstObjectPtr obj,
 			StrID methodID,
 			std::span<const TypeID> argTypeIDs = {},
-			ArgsBuffer args_buffer = nullptr,
+			ArgPtrBuffer argptr_buffer = nullptr,
 			std::pmr::memory_resource* result_rsrc = std::pmr::get_default_resource()) const;
 
 		SharedObject MInvoke(
 			ObjectPtr obj,
 			StrID methodID,
 			std::span<const TypeID> argTypeIDs = {},
-			ArgsBuffer args_buffer = nullptr,
+			ArgPtrBuffer argptr_buffer = nullptr,
 			std::pmr::memory_resource* result_rsrc = std::pmr::get_default_resource()) const;
 
 		template<typename... Args>
@@ -630,8 +630,8 @@ namespace Ubpa::UDRefl {
 			StrID methodID,
 			Args&&... args) const;
 
-		ObjectPtr NonArgCopyMNew(TypeID      typeID, std::pmr::memory_resource* rsrc, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const;
-		ObjectPtr MNew          (TypeID      typeID, std::pmr::memory_resource* rsrc, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const;
+		ObjectPtr NonArgCopyMNew(TypeID      typeID, std::pmr::memory_resource* rsrc, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const;
+		ObjectPtr MNew          (TypeID      typeID, std::pmr::memory_resource* rsrc, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const;
 		bool      MDelete       (ConstObjectPtr obj, std::pmr::memory_resource* rsrc) const;
 
 		template<typename... Args>