+++ src/core/MethodPtr.cpp
@@ -2,27 +2,27 @@
 
 using namespace Ubpa::UDRefl;
 
-Destructor MethodPtr::Invoke(void* obj, void* result_buffer, ArgsBuffer args_buffer) const {
+Destructor MethodPtr::Invoke(void* obj, void* result_buffer, ArgPtrBuffer argptr_buffer) const {
 	return std::visit([=, this](const auto& f) {
 		using Func = std::decay_t<decltype(f)>;
 		if constexpr (std::is_same_v<Func, MemberVariableFunction*>)
-			return f(obj, result_buffer, { args_buffer,paramList });
+			return f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, MemberConstFunction*>)
-			return f(obj, result_buffer, { args_buffer,paramList });
+			return f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, StaticFunction*>)
-			return f(     result_buffer, { args_buffer,paramList });
+			return f(     result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, std::function<MemberVariableFunction>>)
-			return f(obj, result_buffer, { args_buffer,paramList });
+			return f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, std::function<MemberConstFunction>>)
-			return f(obj, result_buffer, { args_buffer,paramList });
+			return f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, std::function<StaticFunction>>)
-			return f(     result_buffer, { args_buffer,paramList });
+			return f(     result_buffer, { argptr_buffer,paramList });
 		else
 			static_assert(always_false<Func>);
 	}, func);
 };
 
-Destructor MethodPtr::Invoke(const void* obj, void* result_buffer, ArgsBuffer args_buffer) const {
+Destructor MethodPtr::Invoke(const void* obj, void* result_buffer, ArgPtrBuffer argptr_buffer) const {
 	return std::visit([=, this](const auto& f)->Destructor {
 		using Func = std::decay_t<decltype(f)>;
 		if constexpr (std::is_same_v<Func, MemberVariableFunction*>) {
@@ -30,23 +30,23 @@ Destructor MethodPtr::Invoke(const void* obj, void* result_buffer, ArgsBuffer ar
 			return {};
 		}
 		else if constexpr (std::is_same_v<Func, MemberConstFunction*>)
-			return f(obj, result_buffer, { args_buffer,paramList });
+			return f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, StaticFunction*>)
-			return f(     result_buffer, { args_buffer,paramList });
+			return f(     result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, std::function<MemberVariableFunction>>) {
 			assert(false);
 			return {};
 		}
 		else if constexpr (std::is_same_v<Func, std::function<MemberConstFunction>>)
-			return f(obj, result_buffer, { args_buffer,paramList });
+			return f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, std::function<StaticFunction>>)
-			return f(     result_buffer, { args_buffer,paramList });
+			return f(     result_buffer, { argptr_buffer,paramList });
 		else
 			static_assert(always_false<Func>);
 	}, func);
 };
 
-Destructor MethodPtr::Invoke(void* result_buffer, ArgsBuffer args_buffer) const {
+Destructor MethodPtr::Invoke(void* result_buffer, ArgPtrBuffer argptr_buffer) const {
 	return std::visit([=, this](const auto& f)->Destructor {
 		using Func = std::decay_t<decltype(f)>;
 		if constexpr (std::is_same_v<Func, MemberVariableFunction*>) {
@@ -58,7 +58,7 @@ Destructor MethodPtr::Invoke(void* result_buffer, ArgsBuffer args_buffer) const 
 			return {};
 		}
 		else if constexpr (std::is_same_v<Func, StaticFunction*>)
-			return f(result_buffer, { args_buffer,paramList });
+			return f(result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, std::function<MemberVariableFunction>>) {
 			assert(false);
 			return {};
@@ -68,7 +68,7 @@ Destructor MethodPtr::Invoke(void* result_buffer, ArgsBuffer args_buffer) const 
 			return {};
 		}
 		else if constexpr (std::is_same_v<Func, std::function<StaticFunction>>)
-			return f(result_buffer, { args_buffer,paramList });
+			return f(result_buffer, { argptr_buffer,paramList });
 		else
 			static_assert(always_false<Func>);
 	}, func);