+++ src/core/ReflMngr.cpp
@@ -182,6 +182,7 @@ namespace Ubpa::UDRefl::details {
 			std::size_t idx;
 			std::size_t typeID; // doesn't use TypeID to avoid init on stack
 		};
+		static_assert(alignof(NonPtrArgInfo) == alignof(void*));
 		struct PtrArgInfo {
 			std::size_t idx;
 			void* ptr;
@@ -191,13 +192,15 @@ namespace Ubpa::UDRefl::details {
 			std::pmr::memory_resource* rsrc,
 			std::span<const TypeID> paramTypeIDs,
 			std::span<const TypeID> argTypeIDs,
-			ArgsBuffer orig_args_buffer) :
+			ArgPtrBuffer orig_argptr_buffer) :
 			rsrc{ rsrc },
 			num_args{ argTypeIDs.size() }
 		{
 			NonPtrArgInfo nonptrargs_stack[MaxArgNum];
+			std::size_t offset_nonptrargs[MaxArgNum];
 			PtrArgInfo ptrargs_stack[MaxArgNum];
 			std::size_t num_ptrargs = 0;
+			std::size_t size_nonptrargs = 0;
 
 			for (std::size_t i = 0; i < argTypeIDs.size(); i++) {
 				if (paramTypeIDs[i] == argTypeIDs[i])
@@ -242,68 +245,84 @@ namespace Ubpa::UDRefl::details {
 
 				std::string_view name = Mngr->tregistry.Nameof(paramTypeIDs[i]);
 				if (type_name_is_pointer(name)) {
-					ptrargs_stack[num_ptrargs++] = { i, orig_args_buffer[i] }; // copy ptr
+					ptrargs_stack[num_ptrargs++] = { i, orig_argptr_buffer[i] }; // copy ptr
 					assert(num_ptrargs <= MaxArgNum);
 				}
 				else {
 					TypeID raw_param_ID = TypeID{ type_name_remove_cvref(lhs) };
-					nonptrargs_stack[num_nonptrargs++] = { i, raw_param_ID.GetValue() };
+					const auto& typeinfo = Mngr->typeinfos.at(raw_param_ID);
+					std::size_t offset = (size_nonptrargs + (typeinfo.alignment - 1)) / typeinfo.alignment * typeinfo.alignment;
+					nonptrargs_stack[num_nonptrargs] = { i, raw_param_ID.GetValue() };
+					offset_nonptrargs[num_nonptrargs] = offset;
+					size_nonptrargs = offset + typeinfo.size;
+					if (typeinfo.alignment > max_alignment)
+						max_alignment = typeinfo.alignment;
+					num_nonptrargs++;
 					assert(num_nonptrargs <= MaxArgNum);
 				}
 			}
 
 			if (num_nonptrargs + num_ptrargs == 0)
-				args_buffer = orig_args_buffer;
+				argptr_buffer = orig_argptr_buffer;
 			else {
-				// buffer = constructed_args_buffer + nonptrargs
-				buffer_size = num_args * sizeof(void*) + num_nonptrargs * sizeof(NonPtrArgInfo);
-				buffer = rsrc->allocate(buffer_size);
-				constructed_args_buffer = reinterpret_cast<void**>(forward_offset(buffer, 0));
-				nonptrargs = reinterpret_cast<NonPtrArgInfo*>(forward_offset(buffer, num_args * sizeof(void*)));
+				// buffer = constructed argptr buffer + nonptrarginfos + copied args
 
-				std::memcpy(constructed_args_buffer, orig_args_buffer, num_args * sizeof(void*));
-				std::memcpy(nonptrargs, nonptrargs_stack, num_nonptrargs * sizeof(NonPtrArgInfo));
+				std::size_t offset_constructed_argptr_buffer = 0;
+				std::size_t offset_nonptrarginfos = num_args * sizeof(void*);
+				std::size_t offset_nonptrargs_buffer =
+					(offset_nonptrarginfos + num_nonptrargs * sizeof(NonPtrArgInfo) + max_alignment - 1) / max_alignment * max_alignment;
+				buffer_size = offset_nonptrargs_buffer + size_nonptrargs;
+
+				buffer = rsrc->allocate(buffer_size, max_alignment);
+				constructed_argptr_buffer = reinterpret_cast<void**>(forward_offset(buffer, offset_constructed_argptr_buffer));
+				nonptrarginfos = reinterpret_cast<NonPtrArgInfo*>(forward_offset(buffer, offset_nonptrarginfos));
+				void* nonptrargs_buffer = forward_offset(buffer, offset_nonptrargs_buffer);
+
+				std::memcpy(constructed_argptr_buffer, orig_argptr_buffer, num_args * sizeof(void*));
+				std::memcpy(nonptrarginfos, nonptrargs_stack, num_nonptrargs * sizeof(NonPtrArgInfo));
 
 				for (std::size_t i = 0; i < num_nonptrargs; i++) {
-					ObjectPtr constructedArg = Mngr->NonArgCopyMNew(
-						TypeID{ nonptrargs_stack[i].typeID },
-						rsrc,
+					void* ptr = forward_offset(nonptrargs_buffer, offset_nonptrargs[i]);
+					bool success = Mngr->NonArgCopyConstruct(
+						ObjectPtr{ TypeID{ nonptrargs_stack[i].typeID }, ptr },
 						std::span<const TypeID>{&argTypeIDs[i], 1},
-						static_cast<ArgsBuffer>(&orig_args_buffer[i])
+						static_cast<ArgPtrBuffer>(&orig_argptr_buffer[i])
 					);
-					constructed_args_buffer[nonptrargs_stack[i].idx] = constructedArg.GetPtr();
+					assert(success);
+					constructed_argptr_buffer[nonptrargs_stack[i].idx] = ptr;
 				}
 
 				for (std::size_t i = 0; i < num_ptrargs; i++)
-					constructed_args_buffer[ptrargs_stack[i].idx] = ptrargs_stack[i].ptr;
+					constructed_argptr_buffer[ptrargs_stack[i].idx] = ptrargs_stack[i].ptr;
 
-				args_buffer = constructed_args_buffer;
+				argptr_buffer = constructed_argptr_buffer;
 			}
 		}
 
 		~ConstructedArgumentsGuard() {
 			if (buffer) {
-				for (const auto& arg : std::span<NonPtrArgInfo>{ nonptrargs, num_nonptrargs }) {
-					bool success = Mngr->MDelete({ TypeID{ arg.typeID }, constructed_args_buffer[arg.idx] }, rsrc);
+				for (const auto& arg : std::span<NonPtrArgInfo>{ nonptrarginfos, num_nonptrargs }) {
+					bool success = Mngr->Destruct({ TypeID{ arg.typeID }, constructed_argptr_buffer[arg.idx] });
 					assert(success);
 				}
-				rsrc->deallocate(buffer, buffer_size);
+				rsrc->deallocate(buffer, buffer_size, max_alignment);
 			}
 		}
 
 		ConstructedArgumentsGuard(const ConstructedArgumentsGuard&) = delete;
 		ConstructedArgumentsGuard& operator=(ConstructedArgumentsGuard&) = delete;
 
-		ArgsBuffer GetArgsBuffer() const noexcept { return args_buffer; }
+		ArgPtrBuffer GetArgPtrBuffer() const noexcept { return argptr_buffer; }
 	private:
 		std::pmr::memory_resource* rsrc;
 		std::size_t buffer_size{ 0 };
+		std::size_t max_alignment{ alignof(std::max_align_t) };
 		void* buffer{ nullptr };
-		void** constructed_args_buffer{ nullptr };
-		NonPtrArgInfo* nonptrargs{ nullptr };
+		void** constructed_argptr_buffer{ nullptr };
+		NonPtrArgInfo* nonptrarginfos{ nullptr };
 		std::size_t num_args;
 		std::size_t num_nonptrargs{ 0 };
-		ArgsBuffer args_buffer{ nullptr };
+		ArgPtrBuffer argptr_buffer{ nullptr };
 	};
 
 	static InvocableResult IsStaticInvocable(
@@ -424,7 +443,7 @@ namespace Ubpa::UDRefl::details {
 		StrID methodID,
 		void* result_buffer,
 		std::span<const TypeID> argTypeIDs,
-		ArgsBuffer args_buffer)
+		ArgPtrBuffer argptr_buffer)
 	{
 		assert(Mngr->GetDereferenceProperty(typeID) == DereferenceProperty::NotReference);
 
@@ -442,17 +461,17 @@ namespace Ubpa::UDRefl::details {
 				&& (is_priority ? IsPriorityCompatible(mtarget->second.methodptr.GetParamList(), argTypeIDs)
 					: Mngr->IsCompatible(mtarget->second.methodptr.GetParamList(), argTypeIDs)))
 			{
-				ConstructedArgumentsGuard guard{ args_rsrc, mtarget->second.methodptr.GetParamList(), argTypeIDs, args_buffer };
+				ConstructedArgumentsGuard guard{ args_rsrc, mtarget->second.methodptr.GetParamList(), argTypeIDs, argptr_buffer };
 				return {
 					true,
 					mtarget->second.methodptr.GetResultDesc().typeID,
-					std::move(mtarget->second.methodptr.Invoke(result_buffer, guard.GetArgsBuffer()))
+					std::move(mtarget->second.methodptr.Invoke(result_buffer, guard.GetArgPtrBuffer()))
 				};
 			}
 		}
 
 		for (const auto& [baseID, baseinfo] : typeinfo.baseinfos) {
-			auto rst = Invoke(is_priority, args_rsrc, baseID, methodID, result_buffer, argTypeIDs, args_buffer);
+			auto rst = Invoke(is_priority, args_rsrc, baseID, methodID, result_buffer, argTypeIDs, argptr_buffer);
 			if (rst.success)
 				return rst;
 		}
@@ -467,7 +486,7 @@ namespace Ubpa::UDRefl::details {
 		StrID methodID,
 		void* result_buffer,
 		std::span<const TypeID> argTypeIDs,
-		ArgsBuffer args_buffer)
+		ArgPtrBuffer argptr_buffer)
 	{
 		assert(Mngr->GetDereferenceProperty(obj.GetID()) == DereferenceProperty::NotReference);
 
@@ -485,11 +504,11 @@ namespace Ubpa::UDRefl::details {
 				&& (is_priority ? IsPriorityCompatible(mtarget->second.methodptr.GetParamList(), argTypeIDs)
 					: Mngr->IsCompatible(mtarget->second.methodptr.GetParamList(), argTypeIDs)))
 			{
-				ConstructedArgumentsGuard guard{ args_rsrc, mtarget->second.methodptr.GetParamList(), argTypeIDs, args_buffer };
+				ConstructedArgumentsGuard guard{ args_rsrc, mtarget->second.methodptr.GetParamList(), argTypeIDs, argptr_buffer };
 				return {
 					true,
 					mtarget->second.methodptr.GetResultDesc().typeID,
-					std::move(mtarget->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgsBuffer()))
+					std::move(mtarget->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgPtrBuffer()))
 				};
 			}
 		}
@@ -498,7 +517,7 @@ namespace Ubpa::UDRefl::details {
 			auto rst = Invoke(
 				is_priority, args_rsrc,
 				ConstObjectPtr{ baseID, baseinfo.StaticCast_DerivedToBase(obj.GetPtr()) },
-				methodID, result_buffer, argTypeIDs, args_buffer
+				methodID, result_buffer, argTypeIDs, argptr_buffer
 			);
 			if (rst.success)
 				return rst;
@@ -514,7 +533,7 @@ namespace Ubpa::UDRefl::details {
 		StrID methodID,
 		void* result_buffer,
 		std::span<const TypeID> argTypeIDs,
-		ArgsBuffer args_buffer)
+		ArgPtrBuffer argptr_buffer)
 	{
 		assert(Mngr->GetDereferenceProperty(obj.GetID()) == DereferenceProperty::NotReference);
 
@@ -535,11 +554,11 @@ namespace Ubpa::UDRefl::details {
 					&& (is_priority ? IsPriorityCompatible(iter->second.methodptr.GetParamList(), argTypeIDs)
 						: Mngr->IsCompatible(iter->second.methodptr.GetParamList(), argTypeIDs)))
 				{
-					ConstructedArgumentsGuard guard{ args_rsrc, iter->second.methodptr.GetParamList(), argTypeIDs, args_buffer };
+					ConstructedArgumentsGuard guard{ args_rsrc, iter->second.methodptr.GetParamList(), argTypeIDs, argptr_buffer };
 					return {
 						true,
 						iter->second.methodptr.GetResultDesc().typeID,
-						iter->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgsBuffer())
+						iter->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgPtrBuffer())
 					};
 				}
 			}
@@ -552,11 +571,11 @@ namespace Ubpa::UDRefl::details {
 					&& (is_priority ? IsPriorityCompatible(iter->second.methodptr.GetParamList(), argTypeIDs)
 						: Mngr->IsCompatible(iter->second.methodptr.GetParamList(), argTypeIDs)))
 				{
-					ConstructedArgumentsGuard guard{ args_rsrc, iter->second.methodptr.GetParamList(), argTypeIDs, args_buffer };
+					ConstructedArgumentsGuard guard{ args_rsrc, iter->second.methodptr.GetParamList(), argTypeIDs, argptr_buffer };
 					return {
 						true,
 						iter->second.methodptr.GetResultDesc().typeID,
-						iter->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgsBuffer())
+						iter->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgPtrBuffer())
 					};
 				}
 			}
@@ -566,7 +585,7 @@ namespace Ubpa::UDRefl::details {
 			auto rst = Invoke(
 				is_priority, args_rsrc,
 				ObjectPtr{ baseID, baseinfo.StaticCast_DerivedToBase(obj.GetPtr()) },
-				methodID, result_buffer, argTypeIDs, args_buffer
+				methodID, result_buffer, argTypeIDs, argptr_buffer
 			);
 			if (rst.success)
 				return rst;
@@ -581,7 +600,7 @@ namespace Ubpa::UDRefl::details {
 		TypeID typeID,
 		StrID methodID,
 		std::span<const TypeID> argTypeIDs,
-		ArgsBuffer args_buffer,
+		ArgPtrBuffer argptr_buffer,
 		std::pmr::memory_resource* rst_rsrc)
 	{
 		assert(Mngr->GetDereferenceProperty(typeID) == DereferenceProperty::NotReference);
@@ -602,10 +621,10 @@ namespace Ubpa::UDRefl::details {
 			{
 				const auto& methodptr = mtarget->second.methodptr;
 				const auto& rst_desc = methodptr.GetResultDesc();
-				ConstructedArgumentsGuard guard{ args_rsrc, mtarget->second.methodptr.GetParamList(), argTypeIDs, args_buffer };
+				ConstructedArgumentsGuard guard{ args_rsrc, mtarget->second.methodptr.GetParamList(), argTypeIDs, argptr_buffer };
 
 				if (rst_desc.IsVoid()) {
-					mtarget->second.methodptr.Invoke(nullptr, guard.GetArgsBuffer());
+					mtarget->second.methodptr.Invoke(nullptr, guard.GetArgPtrBuffer());
 					return {
 						{rst_desc.typeID, nullptr},
 						[](void* ptr) { assert(ptr); }
@@ -613,7 +632,7 @@ namespace Ubpa::UDRefl::details {
 				}
 				else if (type_name_is_reference(Mngr->tregistry.Nameof(rst_desc.typeID))) {
 					std::aligned_storage_t<sizeof(void*)> buffer;
-					mtarget->second.methodptr.Invoke(&buffer, guard.GetArgsBuffer());
+					mtarget->second.methodptr.Invoke(&buffer, guard.GetArgPtrBuffer());
 					return {
 						{rst_desc.typeID, buffer_as<void*>(&buffer)},
 						[](void* ptr) { assert(ptr); }
@@ -621,7 +640,7 @@ namespace Ubpa::UDRefl::details {
 				}
 				else {
 					void* result_buffer = rst_rsrc->allocate(rst_desc.size, rst_desc.alignment);
-					auto dtor = mtarget->second.methodptr.Invoke(result_buffer, guard.GetArgsBuffer());
+					auto dtor = mtarget->second.methodptr.Invoke(result_buffer, guard.GetArgPtrBuffer());
 					return {
 						{rst_desc.typeID, result_buffer},
 						GenerateDeleteFunc(std::move(dtor), rst_rsrc, rst_desc.size, rst_desc.alignment)
@@ -631,7 +650,7 @@ namespace Ubpa::UDRefl::details {
 		}
 
 		for (const auto& [baseID, baseinfo] : typeinfo.baseinfos) {
-			auto rst = MInvoke(is_priority, args_rsrc, baseID, methodID, argTypeIDs, args_buffer, rst_rsrc);
+			auto rst = MInvoke(is_priority, args_rsrc, baseID, methodID, argTypeIDs, argptr_buffer, rst_rsrc);
 			if (rst.GetID())
 				return rst;
 		}
@@ -645,7 +664,7 @@ namespace Ubpa::UDRefl::details {
 		ConstObjectPtr obj,
 		StrID methodID,
 		std::span<const TypeID> argTypeIDs,
-		ArgsBuffer args_buffer,
+		ArgPtrBuffer argptr_buffer,
 		std::pmr::memory_resource* rst_rsrc)
 	{
 		assert(rst_rsrc);
@@ -667,9 +686,9 @@ namespace Ubpa::UDRefl::details {
 			{
 				const auto& methodptr = mtarget->second.methodptr;
 				const auto& rst_desc = methodptr.GetResultDesc();
-				ConstructedArgumentsGuard guard{ args_rsrc, mtarget->second.methodptr.GetParamList(), argTypeIDs, args_buffer };
+				ConstructedArgumentsGuard guard{ args_rsrc, mtarget->second.methodptr.GetParamList(), argTypeIDs, argptr_buffer };
 				if (rst_desc.IsVoid()) {
-					auto dtor = mtarget->second.methodptr.Invoke(obj.GetPtr(), nullptr, guard.GetArgsBuffer());
+					auto dtor = mtarget->second.methodptr.Invoke(obj.GetPtr(), nullptr, guard.GetArgPtrBuffer());
 					return {
 						{rst_desc.typeID, nullptr},
 						[](void* ptr) { assert(!ptr); }
@@ -677,7 +696,7 @@ namespace Ubpa::UDRefl::details {
 				}
 				else if (type_name_is_reference(Mngr->tregistry.Nameof(rst_desc.typeID))) {
 					std::aligned_storage_t<sizeof(void*)> buffer;
-					mtarget->second.methodptr.Invoke(obj.GetPtr(), &buffer, guard.GetArgsBuffer());
+					mtarget->second.methodptr.Invoke(obj.GetPtr(), &buffer, guard.GetArgPtrBuffer());
 					return {
 						{rst_desc.typeID, buffer_as<void*>(&buffer)},
 						[](void* ptr) { assert(ptr); }
@@ -685,7 +704,7 @@ namespace Ubpa::UDRefl::details {
 				}
 				else {
 					void* result_buffer = rst_rsrc->allocate(rst_desc.size, rst_desc.alignment);
-					auto dtor = mtarget->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgsBuffer());
+					auto dtor = mtarget->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgPtrBuffer());
 					return {
 						{rst_desc.typeID, result_buffer},
 						GenerateDeleteFunc(std::move(dtor), rst_rsrc, rst_desc.size, rst_desc.alignment)
@@ -698,7 +717,7 @@ namespace Ubpa::UDRefl::details {
 			auto rst = MInvoke(
 				is_priority, args_rsrc,
 				ConstObjectPtr{ baseID, baseinfo.StaticCast_DerivedToBase(obj.GetPtr()) },
-				methodID, argTypeIDs, args_buffer, rst_rsrc
+				methodID, argTypeIDs, argptr_buffer, rst_rsrc
 			);
 			if (rst.GetID())
 				return rst;
@@ -713,7 +732,7 @@ namespace Ubpa::UDRefl::details {
 		ObjectPtr obj,
 		StrID methodID,
 		std::span<const TypeID> argTypeIDs,
-		ArgsBuffer args_buffer,
+		ArgPtrBuffer argptr_buffer,
 		std::pmr::memory_resource* rst_rsrc)
 	{
 		assert(rst_rsrc);
@@ -738,10 +757,10 @@ namespace Ubpa::UDRefl::details {
 				{
 					const auto& methodptr = iter->second.methodptr;
 					const auto& rst_desc = methodptr.GetResultDesc();
-					ConstructedArgumentsGuard guard{ args_rsrc, iter->second.methodptr.GetParamList(), argTypeIDs, args_buffer };
+					ConstructedArgumentsGuard guard{ args_rsrc, iter->second.methodptr.GetParamList(), argTypeIDs, argptr_buffer };
 
 					if (rst_desc.IsVoid()) {
-						auto dtor = iter->second.methodptr.Invoke(obj.GetPtr(), nullptr, guard.GetArgsBuffer());
+						auto dtor = iter->second.methodptr.Invoke(obj.GetPtr(), nullptr, guard.GetArgPtrBuffer());
 						return {
 							{rst_desc.typeID, nullptr},
 							[](void* ptr) { assert(!ptr); }
@@ -749,7 +768,7 @@ namespace Ubpa::UDRefl::details {
 					}
 					else if (type_name_is_reference(Mngr->tregistry.Nameof(rst_desc.typeID))) {
 						std::aligned_storage_t<sizeof(void*)> buffer;
-						iter->second.methodptr.Invoke(obj.GetPtr(), &buffer, guard.GetArgsBuffer());
+						iter->second.methodptr.Invoke(obj.GetPtr(), &buffer, guard.GetArgPtrBuffer());
 						return {
 							{rst_desc.typeID, buffer_as<void*>(&buffer)},
 							[](void* ptr) { assert(ptr); }
@@ -757,7 +776,7 @@ namespace Ubpa::UDRefl::details {
 					}
 					else {
 						void* result_buffer = rst_rsrc->allocate(rst_desc.size, rst_desc.alignment);
-						auto dtor = iter->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgsBuffer());
+						auto dtor = iter->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgPtrBuffer());
 						return {
 							{rst_desc.typeID, result_buffer},
 							GenerateDeleteFunc(std::move(dtor), rst_rsrc, rst_desc.size, rst_desc.alignment)
@@ -776,9 +795,9 @@ namespace Ubpa::UDRefl::details {
 				{
 					const auto& methodptr = iter->second.methodptr;
 					const auto& rst_desc = methodptr.GetResultDesc();
-					ConstructedArgumentsGuard guard{ args_rsrc, iter->second.methodptr.GetParamList(), argTypeIDs, args_buffer };
+					ConstructedArgumentsGuard guard{ args_rsrc, iter->second.methodptr.GetParamList(), argTypeIDs, argptr_buffer };
 					if (rst_desc.IsVoid()) {
-						auto dtor = iter->second.methodptr.Invoke(obj.GetPtr(), nullptr, guard.GetArgsBuffer());
+						auto dtor = iter->second.methodptr.Invoke(obj.GetPtr(), nullptr, guard.GetArgPtrBuffer());
 						return {
 							{rst_desc.typeID, nullptr},
 							[](void* ptr) { assert(!ptr); }
@@ -786,7 +805,7 @@ namespace Ubpa::UDRefl::details {
 					}
 					else if (type_name_is_reference(Mngr->tregistry.Nameof(rst_desc.typeID))) {
 						std::aligned_storage_t<sizeof(void*)> buffer;
-						iter->second.methodptr.Invoke(obj.GetPtr(), &buffer, guard.GetArgsBuffer());
+						iter->second.methodptr.Invoke(obj.GetPtr(), &buffer, guard.GetArgPtrBuffer());
 						return {
 							{rst_desc.typeID, buffer_as<void*>(&buffer)},
 							[](void* ptr) { assert(ptr); }
@@ -794,7 +813,7 @@ namespace Ubpa::UDRefl::details {
 					}
 					else {
 						void* result_buffer = rst_rsrc->allocate(rst_desc.size, rst_desc.alignment);
-						auto dtor = iter->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgsBuffer());
+						auto dtor = iter->second.methodptr.Invoke(obj.GetPtr(), result_buffer, guard.GetArgPtrBuffer());
 						return {
 							{rst_desc.typeID, result_buffer},
 							GenerateDeleteFunc(std::move(dtor), rst_rsrc, rst_desc.size, rst_desc.alignment)
@@ -808,7 +827,7 @@ namespace Ubpa::UDRefl::details {
 			auto rst = MInvoke(
 				is_priority, args_rsrc,
 				ObjectPtr{ baseID, baseinfo.StaticCast_DerivedToBase(obj.GetPtr()) },
-				methodID, argTypeIDs, args_buffer, rst_rsrc
+				methodID, argTypeIDs, argptr_buffer, rst_rsrc
 			);
 			if (rst.GetID())
 				return rst;
@@ -1207,7 +1226,7 @@ bool ReflMngr::AlignedFree(void* ptr) const {
 	return InvokeArgs(GlobalID, StrIDRegistry::MetaID::aligned_free, nullptr, std::move(ptr));
 }
 
-ObjectPtr ReflMngr::NonArgCopyNew(TypeID typeID, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const {
+ObjectPtr ReflMngr::NonArgCopyNew(TypeID typeID, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const {
 	if (!IsNonArgCopyConstructible(typeID, argTypeIDs))
 		return nullptr;
 
@@ -1221,13 +1240,13 @@ ObjectPtr ReflMngr::NonArgCopyNew(TypeID typeID, std::span<const TypeID> argType
 		return nullptr;
 
 	ObjectPtr obj{ typeID, buffer };
-	bool success = NonArgCopyConstruct(obj, argTypeIDs, args_buffer);
+	bool success = NonArgCopyConstruct(obj, argTypeIDs, argptr_buffer);
 	assert(success);
 
 	return obj;
 }
 
-ObjectPtr ReflMngr::New(TypeID typeID, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const {
+ObjectPtr ReflMngr::New(TypeID typeID, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const {
 	if (!IsConstructible(typeID, argTypeIDs))
 		return nullptr;
 
@@ -1241,7 +1260,7 @@ ObjectPtr ReflMngr::New(TypeID typeID, std::span<const TypeID> argTypeIDs, ArgsB
 		return nullptr;
 
 	ObjectPtr obj{ typeID, buffer };
-	bool success = Construct(obj, argTypeIDs, args_buffer);
+	bool success = Construct(obj, argTypeIDs, argptr_buffer);
 	assert(success);
 	
 	return obj;
@@ -1263,8 +1282,8 @@ bool ReflMngr::Delete(ConstObjectPtr obj) const {
 	return free_success;
 }
 
-SharedObject ReflMngr::MakeShared(TypeID typeID, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const {
-	ObjectPtr obj = New(typeID, argTypeIDs, args_buffer);
+SharedObject ReflMngr::MakeShared(TypeID typeID, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const {
+	ObjectPtr obj = New(typeID, argTypeIDs, argptr_buffer);
 
 	if (!obj.Valid())
 		return { obj, [](void*) {} };
@@ -1653,15 +1672,15 @@ InvokeResult ReflMngr::Invoke(
 	StrID methodID,
 	void* result_buffer,
 	std::span<const TypeID> argTypeIDs,
-	ArgsBuffer args_buffer) const
+	ArgPtrBuffer argptr_buffer) const
 {
 	if (GetDereferenceProperty(typeID) != DereferenceProperty::NotReference)
-		return Invoke(Dereference(typeID), methodID, result_buffer, argTypeIDs, args_buffer);
+		return Invoke(Dereference(typeID), methodID, result_buffer, argTypeIDs, argptr_buffer);
 
-	if(auto priority_rst = details::Invoke(true, &temporary_resource, typeID, methodID, result_buffer, argTypeIDs, args_buffer))
+	if(auto priority_rst = details::Invoke(true, &temporary_resource, typeID, methodID, result_buffer, argTypeIDs, argptr_buffer))
 		return priority_rst;
 
-	return details::Invoke(false, &temporary_resource, typeID, methodID, result_buffer, argTypeIDs, args_buffer);
+	return details::Invoke(false, &temporary_resource, typeID, methodID, result_buffer, argTypeIDs, argptr_buffer);
 }
 
 InvokeResult ReflMngr::Invoke(
@@ -1669,23 +1688,23 @@ InvokeResult ReflMngr::Invoke(
 	StrID methodID,
 	void* result_buffer,
 	std::span<const TypeID> argTypeIDs,
-	ArgsBuffer args_buffer) const
+	ArgPtrBuffer argptr_buffer) const
 {
 	auto deref_prop = GetDereferenceProperty(obj.GetID());
 	switch (deref_prop)
 	{
 	case Ubpa::UDRefl::DereferenceProperty::Variable:
-		return Invoke(Dereference(obj), methodID, result_buffer, argTypeIDs, args_buffer);
+		return Invoke(Dereference(obj), methodID, result_buffer, argTypeIDs, argptr_buffer);
 	case Ubpa::UDRefl::DereferenceProperty::Const:
-		return Invoke(DereferenceAsConst(obj), methodID, result_buffer, argTypeIDs, args_buffer);
+		return Invoke(DereferenceAsConst(obj), methodID, result_buffer, argTypeIDs, argptr_buffer);
 	default:
 		break;
 	}
 
-	if (auto priority_rst = details::Invoke(true, &temporary_resource, obj, methodID, result_buffer, argTypeIDs, args_buffer))
+	if (auto priority_rst = details::Invoke(true, &temporary_resource, obj, methodID, result_buffer, argTypeIDs, argptr_buffer))
 		return priority_rst;
 
-	return details::Invoke(false, &temporary_resource, obj, methodID, result_buffer, argTypeIDs, args_buffer);
+	return details::Invoke(false, &temporary_resource, obj, methodID, result_buffer, argTypeIDs, argptr_buffer);
 }
 
 InvokeResult ReflMngr::Invoke(
@@ -1693,47 +1712,47 @@ InvokeResult ReflMngr::Invoke(
 	StrID methodID,
 	void* result_buffer,
 	std::span<const TypeID> argTypeIDs,
-	ArgsBuffer args_buffer) const
+	ArgPtrBuffer argptr_buffer) const
 {
 	auto deref_prop = GetDereferenceProperty(obj.GetID());
 	switch (deref_prop)
 	{
 	case Ubpa::UDRefl::DereferenceProperty::Variable:
-		return Invoke(Dereference(obj), methodID, result_buffer, argTypeIDs, args_buffer);
+		return Invoke(Dereference(obj), methodID, result_buffer, argTypeIDs, argptr_buffer);
 	case Ubpa::UDRefl::DereferenceProperty::Const:
-		return Invoke(DereferenceAsConst(obj), methodID, result_buffer, argTypeIDs, args_buffer);
+		return Invoke(DereferenceAsConst(obj), methodID, result_buffer, argTypeIDs, argptr_buffer);
 	default:
 		break;
 	}
 
-	if (auto priority_rst = details::Invoke(true, &temporary_resource, obj, methodID, result_buffer, argTypeIDs, args_buffer))
+	if (auto priority_rst = details::Invoke(true, &temporary_resource, obj, methodID, result_buffer, argTypeIDs, argptr_buffer))
 		return priority_rst;
 
-	return details::Invoke(false, &temporary_resource, obj, methodID, result_buffer, argTypeIDs, args_buffer);
+	return details::Invoke(false, &temporary_resource, obj, methodID, result_buffer, argTypeIDs, argptr_buffer);
 }
 
 SharedObject ReflMngr::MInvoke(
 	TypeID typeID,
 	StrID methodID,
 	std::span<const TypeID> argTypeIDs,
-	ArgsBuffer args_buffer,
+	ArgPtrBuffer argptr_buffer,
 	std::pmr::memory_resource* rst_rsrc) const
 {
 	assert(rst_rsrc);
 	if (GetDereferenceProperty(typeID) != DereferenceProperty::NotReference)
-		return MInvoke(Dereference(typeID), methodID, argTypeIDs, args_buffer, rst_rsrc);
+		return MInvoke(Dereference(typeID), methodID, argTypeIDs, argptr_buffer, rst_rsrc);
 
-	if (auto priority_rst = details::MInvoke(true, &temporary_resource, typeID, methodID, argTypeIDs, args_buffer, rst_rsrc))
+	if (auto priority_rst = details::MInvoke(true, &temporary_resource, typeID, methodID, argTypeIDs, argptr_buffer, rst_rsrc))
 		return priority_rst;
 
-	return details::MInvoke(false, &temporary_resource, typeID, methodID, argTypeIDs, args_buffer, rst_rsrc);
+	return details::MInvoke(false, &temporary_resource, typeID, methodID, argTypeIDs, argptr_buffer, rst_rsrc);
 }
 
 SharedObject ReflMngr::MInvoke(
 	ConstObjectPtr obj,
 	StrID methodID,
 	std::span<const TypeID> argTypeIDs,
-	ArgsBuffer args_buffer,
+	ArgPtrBuffer argptr_buffer,
 	std::pmr::memory_resource* rst_rsrc) const
 {
 	assert(rst_rsrc);
@@ -1741,24 +1760,24 @@ SharedObject ReflMngr::MInvoke(
 	switch (deref_prop)
 	{
 	case Ubpa::UDRefl::DereferenceProperty::Variable:
-		return MInvoke(Dereference(obj), methodID, argTypeIDs, args_buffer, rst_rsrc);
+		return MInvoke(Dereference(obj), methodID, argTypeIDs, argptr_buffer, rst_rsrc);
 	case Ubpa::UDRefl::DereferenceProperty::Const:
-		return MInvoke(DereferenceAsConst(obj), methodID, argTypeIDs, args_buffer, rst_rsrc);
+		return MInvoke(DereferenceAsConst(obj), methodID, argTypeIDs, argptr_buffer, rst_rsrc);
 	default:
 		break;
 	}
 
-	if (auto priority_rst = details::MInvoke(true, &temporary_resource, obj, methodID, argTypeIDs, args_buffer, rst_rsrc))
+	if (auto priority_rst = details::MInvoke(true, &temporary_resource, obj, methodID, argTypeIDs, argptr_buffer, rst_rsrc))
 		return priority_rst;
 
-	return details::MInvoke(false, &temporary_resource, obj, methodID, argTypeIDs, args_buffer, rst_rsrc);
+	return details::MInvoke(false, &temporary_resource, obj, methodID, argTypeIDs, argptr_buffer, rst_rsrc);
 }
 
 SharedObject ReflMngr::MInvoke(
 	ObjectPtr obj,
 	StrID methodID,
 	std::span<const TypeID> argTypeIDs,
-	ArgsBuffer args_buffer,
+	ArgPtrBuffer argptr_buffer,
 	std::pmr::memory_resource* rst_rsrc) const
 {
 	assert(rst_rsrc);
@@ -1766,20 +1785,20 @@ SharedObject ReflMngr::MInvoke(
 	switch (deref_prop)
 	{
 	case Ubpa::UDRefl::DereferenceProperty::Variable:
-		return MInvoke(Dereference(obj), methodID, argTypeIDs, args_buffer, rst_rsrc);
+		return MInvoke(Dereference(obj), methodID, argTypeIDs, argptr_buffer, rst_rsrc);
 	case Ubpa::UDRefl::DereferenceProperty::Const:
-		return MInvoke(DereferenceAsConst(obj), methodID, argTypeIDs, args_buffer, rst_rsrc);
+		return MInvoke(DereferenceAsConst(obj), methodID, argTypeIDs, argptr_buffer, rst_rsrc);
 	default:
 		break;
 	}
 
-	if (auto priority_rst = details::MInvoke(true, &temporary_resource, obj, methodID, argTypeIDs, args_buffer, rst_rsrc))
+	if (auto priority_rst = details::MInvoke(true, &temporary_resource, obj, methodID, argTypeIDs, argptr_buffer, rst_rsrc))
 		return priority_rst;
 
-	return details::MInvoke(false, &temporary_resource, obj, methodID, argTypeIDs, args_buffer, rst_rsrc);
+	return details::MInvoke(false, &temporary_resource, obj, methodID, argTypeIDs, argptr_buffer, rst_rsrc);
 }
 
-ObjectPtr ReflMngr::NonArgCopyMNew(TypeID typeID, std::pmr::memory_resource* rsrc, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const {
+ObjectPtr ReflMngr::NonArgCopyMNew(TypeID typeID, std::pmr::memory_resource* rsrc, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const {
 	assert(rsrc);
 
 	if (!IsConstructible(typeID, argTypeIDs))
@@ -1793,13 +1812,13 @@ ObjectPtr ReflMngr::NonArgCopyMNew(TypeID typeID, std::pmr::memory_resource* rsr
 		return nullptr;
 
 	ObjectPtr obj{ typeID, buffer };
-	bool success = NonArgCopyConstruct(obj, argTypeIDs, args_buffer);
+	bool success = NonArgCopyConstruct(obj, argTypeIDs, argptr_buffer);
 	assert(success);
 
 	return obj;
 }
 
-ObjectPtr ReflMngr::MNew(TypeID typeID, std::pmr::memory_resource* rsrc, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const {
+ObjectPtr ReflMngr::MNew(TypeID typeID, std::pmr::memory_resource* rsrc, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const {
 	assert(rsrc);
 
 	if (!IsConstructible(typeID, argTypeIDs))
@@ -1813,7 +1832,7 @@ ObjectPtr ReflMngr::MNew(TypeID typeID, std::pmr::memory_resource* rsrc, std::sp
 		return nullptr;
 
 	ObjectPtr obj{ typeID, buffer };
-	bool success = Construct(obj, argTypeIDs, args_buffer);
+	bool success = Construct(obj, argTypeIDs, argptr_buffer);
 	assert(success);
 
 	return obj;
@@ -1896,7 +1915,7 @@ bool ReflMngr::IsDestructible(TypeID typeID) const {
 	return false;
 }
 
-bool ReflMngr::NonArgCopyConstruct(ObjectPtr obj, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const {
+bool ReflMngr::NonArgCopyConstruct(ObjectPtr obj, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const {
 	assert(GetDereferenceProperty(obj.GetID()) == DereferenceProperty::NotReference);
 
 	if (!obj.Valid())
@@ -1913,14 +1932,14 @@ bool ReflMngr::NonArgCopyConstruct(ObjectPtr obj, std::span<const TypeID> argTyp
 		if (mtarget->second.methodptr.IsMemberVariable()
 			&& details::IsNonArgCopyConstructCompatible(mtarget->second.methodptr.GetParamList(), argTypeIDs))
 		{
-			mtarget->second.methodptr.Invoke(obj.GetPtr(), nullptr, args_buffer);
+			mtarget->second.methodptr.Invoke(obj.GetPtr(), nullptr, argptr_buffer);
 			return true;
 		}
 	}
 	return false;
 }
 
-bool ReflMngr::Construct(ObjectPtr obj, std::span<const TypeID> argTypeIDs, ArgsBuffer args_buffer) const {
+bool ReflMngr::Construct(ObjectPtr obj, std::span<const TypeID> argTypeIDs, ArgPtrBuffer argptr_buffer) const {
 	assert(GetDereferenceProperty(obj.GetID()) == DereferenceProperty::NotReference);
 
 	if (!obj.Valid())
@@ -1937,8 +1956,8 @@ bool ReflMngr::Construct(ObjectPtr obj, std::span<const TypeID> argTypeIDs, Args
 		if (mtarget->second.methodptr.IsMemberVariable()
 			&& IsCompatible(mtarget->second.methodptr.GetParamList(), argTypeIDs))
 		{
-			details::ConstructedArgumentsGuard guard{ &temporary_resource, mtarget->second.methodptr.GetParamList(), argTypeIDs, args_buffer };
-			mtarget->second.methodptr.Invoke(obj.GetPtr(), nullptr, guard.GetArgsBuffer());
+			details::ConstructedArgumentsGuard guard{ &temporary_resource, mtarget->second.methodptr.GetParamList(), argTypeIDs, argptr_buffer };
+			mtarget->second.methodptr.Invoke(obj.GetPtr(), nullptr, guard.GetArgPtrBuffer());
 			return true;
 		}
 	}